\name{MAplot}
\alias{MAplot}
\title{Moving Average plot}
\usage{
MAplot()
}
\description{
Calculate the moving average of data, then plot the data and line.
}
\examples{
MAplot(date,price,num_ma)
}
\arguments{
  \item{date}{the date of price}
  \item{price}{Daily price}
  \item{num_ma}{the time you want to build}
}
